requirements.txt
numpy==1.26.4
numpy_stl==3.1.1
pymadcad==0.16.0
